requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
coverage==7.4.0
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==3.2.23
django-debug-toolbar==4.3.0
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-spectacular==0.27.2
Faker==12.0.1
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mixer==7.2.2
oauthlib==3.2.2
packaging==24.0
pillow==10.2.0
pluggy==1.4.0
py==1.11.0
pycparser==2.22
PyJWT==2.8.0
pytest==6.2.5
pytest-django==4.5.2
pytest-pythonpath==0.7.4
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.18.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sorl-thumbnail==12.10.0
sqlparse==0.4.4
toml==0.10.2
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
